refactor: tighten skill-eval sweep structure
