+++ src/benchflow/_skill_eval_constants.py
@@ -1,0 +1,9 @@
+"""Shared constants for skill evaluation."""
+
+DEFAULT_SKILL_MOUNT_DIR = "/skills"
+JUDGE_API_ENV_KEYS = (
+    "GOOGLE_API_KEY",
+    "GEMINI_API_KEY",
+    "ANTHROPIC_API_KEY",
+    "OPENAI_API_KEY",
+)
